+++ google/cloud/asset_v1/types/assets/responses.py
@@ -1,0 +1,31 @@
+# -*- coding: utf-8 -*-
+# Copyright 2022 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+import proto  # type: ignore
+
+from google.cloud.orgpolicy.v1 import orgpolicy_pb2  # type: ignore
+from google.cloud.osconfig_v1 import Inventory  # type: ignore
+from google.iam.v1 import policy_pb2  # type: ignore
+from google.identity.accesscontextmanager.v1 import access_level_pb2  # type: ignore
+from google.identity.accesscontextmanager.v1 import access_policy_pb2  # type: ignore
+from google.identity.accesscontextmanager.v1 import service_perimeter_pb2  # type: ignore
+from google.protobuf import struct_pb2  # type: ignore
+from google.protobuf import timestamp_pb2  # type: ignore
+from google.rpc import code_pb2  # type: ignore
+
+__manifest__ = ()
+
+
+__all__ = tuple(sorted(__manifest__))
