+++ google/cloud/asset_v1p2beta1/types/asset_service/requests.py
@@ -1,0 +1,155 @@
+# -*- coding: utf-8 -*-
+# Copyright 2022 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+import proto  # type: ignore
+
+from google.protobuf import field_mask_pb2  # type: ignore
+
+__manifest__ = (
+    "CreateFeedRequest",
+    "GetFeedRequest",
+    "ListFeedsRequest",
+    "UpdateFeedRequest",
+    "DeleteFeedRequest",
+)
+
+
+class CreateFeedRequest(proto.Message):
+    r"""Create asset feed request.
+
+    Attributes:
+        parent (str):
+            Required. The name of the
+            project/folder/organization where this feed
+            should be created in. It can only be an
+            organization number (such as
+            "organizations/123"), a folder number (such as
+            "folders/123"), a project ID (such as
+            "projects/my-project-id")", or a project number
+            (such as "projects/12345").
+        feed_id (str):
+            Required. This is the client-assigned asset
+            feed identifier and it needs to be unique under
+            a specific parent project/folder/organization.
+        feed (google.cloud.asset_v1p2beta1.types.Feed):
+            Required. The feed details. The field ``name`` must be empty
+            and it will be generated in the format of:
+            projects/project_number/feeds/feed_id
+            folders/folder_number/feeds/feed_id
+            organizations/organization_number/feeds/feed_id
+    """
+    __module__ = __module__.rsplit(".", maxsplit=1)[0]  # type: ignore
+
+    parent = proto.Field(
+        proto.STRING,
+        number=1,
+    )
+    feed_id = proto.Field(
+        proto.STRING,
+        number=2,
+    )
+    feed = proto.Field(
+        proto.MESSAGE,
+        number=3,
+        message="Feed",
+    )
+
+
+class GetFeedRequest(proto.Message):
+    r"""Get asset feed request.
+
+    Attributes:
+        name (str):
+            Required. The name of the Feed and it must be in the format
+            of: projects/project_number/feeds/feed_id
+            folders/folder_number/feeds/feed_id
+            organizations/organization_number/feeds/feed_id
+    """
+    __module__ = __module__.rsplit(".", maxsplit=1)[0]  # type: ignore
+
+    name = proto.Field(
+        proto.STRING,
+        number=1,
+    )
+
+
+class ListFeedsRequest(proto.Message):
+    r"""List asset feeds request.
+
+    Attributes:
+        parent (str):
+            Required. The parent
+            project/folder/organization whose feeds are to
+            be listed. It can only be using
+            project/folder/organization number (such as
+            "folders/12345")", or a project ID (such as
+            "projects/my-project-id").
+    """
+    __module__ = __module__.rsplit(".", maxsplit=1)[0]  # type: ignore
+
+    parent = proto.Field(
+        proto.STRING,
+        number=1,
+    )
+
+
+class UpdateFeedRequest(proto.Message):
+    r"""Update asset feed request.
+
+    Attributes:
+        feed (google.cloud.asset_v1p2beta1.types.Feed):
+            Required. The new values of feed details. It must match an
+            existing feed and the field ``name`` must be in the format
+            of: projects/project_number/feeds/feed_id or
+            folders/folder_number/feeds/feed_id or
+            organizations/organization_number/feeds/feed_id.
+        update_mask (google.protobuf.field_mask_pb2.FieldMask):
+            Required. Only updates the ``feed`` fields indicated by this
+            mask. The field mask must not be empty, and it must not
+            contain fields that are immutable or only set by the server.
+    """
+    __module__ = __module__.rsplit(".", maxsplit=1)[0]  # type: ignore
+
+    feed = proto.Field(
+        proto.MESSAGE,
+        number=1,
+        message="Feed",
+    )
+    update_mask = proto.Field(
+        proto.MESSAGE,
+        number=2,
+        message=field_mask_pb2.FieldMask,
+    )
+
+
+class DeleteFeedRequest(proto.Message):
+    r"""
+
+    Attributes:
+        name (str):
+            Required. The name of the feed and it must be in the format
+            of: projects/project_number/feeds/feed_id
+            folders/folder_number/feeds/feed_id
+            organizations/organization_number/feeds/feed_id
+    """
+    __module__ = __module__.rsplit(".", maxsplit=1)[0]  # type: ignore
+
+    name = proto.Field(
+        proto.STRING,
+        number=1,
+    )
+
+
+__all__ = tuple(sorted(__manifest__))
