+++ google/cloud/asset_v1p2beta1/types/asset_service/responses.py
@@ -1,0 +1,39 @@
+# -*- coding: utf-8 -*-
+# Copyright 2022 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+import proto  # type: ignore
+
+from google.protobuf import field_mask_pb2  # type: ignore
+
+__manifest__ = ("ListFeedsResponse",)
+
+
+class ListFeedsResponse(proto.Message):
+    r"""
+
+    Attributes:
+        feeds (Sequence[google.cloud.asset_v1p2beta1.types.Feed]):
+            A list of feeds.
+    """
+    __module__ = __module__.rsplit(".", maxsplit=1)[0]  # type: ignore
+
+    feeds = proto.RepeatedField(
+        proto.MESSAGE,
+        number=1,
+        message="Feed",
+    )
+
+
+__all__ = tuple(sorted(__manifest__))
