+++ google/cloud/asset_v1p5beta1/types/asset_service/__init__.py
@@ -1,0 +1,50 @@
+# -*- coding: utf-8 -*-
+# Copyright 2022 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+import proto  # type: ignore
+
+from google.cloud.asset_v1p5beta1.types import assets as gca_assets
+from google.protobuf import timestamp_pb2  # type: ignore
+
+
+__protobuf__ = proto.module(
+    package="google.cloud.asset.v1p5beta1",
+    manifest={
+        "ContentType",
+        "ListAssetsRequest",
+        "ListAssetsResponse",
+    },
+)
+
+
+from .requests import (
+    ListAssetsRequest,
+)
+
+from .responses import (
+    ListAssetsResponse,
+)
+
+
+class ContentType(proto.Enum):
+    r"""Asset content type."""
+    CONTENT_TYPE_UNSPECIFIED = 0
+    RESOURCE = 1
+    IAM_POLICY = 2
+    ORG_POLICY = 4
+    ACCESS_POLICY = 5
+
+
+__all__ = tuple(sorted(__protobuf__.manifest))
