+++ google/cloud/asset_v1p5beta1/types/asset_service/responses.py
@@ -1,0 +1,58 @@
+# -*- coding: utf-8 -*-
+# Copyright 2022 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+import proto  # type: ignore
+
+from google.cloud.asset_v1p5beta1.types import assets as gca_assets
+from google.protobuf import timestamp_pb2  # type: ignore
+
+__manifest__ = ("ListAssetsResponse",)
+
+
+class ListAssetsResponse(proto.Message):
+    r"""ListAssets response.
+
+    Attributes:
+        read_time (google.protobuf.timestamp_pb2.Timestamp):
+            Time the snapshot was taken.
+        assets (Sequence[google.cloud.asset_v1p5beta1.types.Asset]):
+            Assets.
+        next_page_token (str):
+            Token to retrieve the next page of results.
+            Set to empty if there are no remaining results.
+    """
+    __module__ = __module__.rsplit(".", maxsplit=1)[0]  # type: ignore
+
+    @property
+    def raw_page(self):
+        return self
+
+    read_time = proto.Field(
+        proto.MESSAGE,
+        number=1,
+        message=timestamp_pb2.Timestamp,
+    )
+    assets = proto.RepeatedField(
+        proto.MESSAGE,
+        number=2,
+        message=gca_assets.Asset,
+    )
+    next_page_token = proto.Field(
+        proto.STRING,
+        number=3,
+    )
+
+
+__all__ = tuple(sorted(__manifest__))
